finalize the player change function and inputs
inputs now are disabled once a move is made and enabled after the
the game is over or restarted. the player change function has also
but finalized so that it changes current player instead of just x
or o. additionally the ability to change a player sign and have
the function attach the correct sign to the gameboard has been added