stop double input when bot goes first
there was an onclick event on the gameBoard that would call the AI
function if player1 was a bot. this was intended to begin the game
and let the computer know to pick a move. however this would get
triggered even on player2's turn which would then cause the program to
generate two inputs on the computer's turn. a check has been added
whereby the gameBoard must be empty to call the AI program